BPM spinner 를 소숫점으로도 처리할 수 있게 수정.
defualt 는 정수 처럼 처리하도록 하지만, 84.5bpm 과 같이 float 데이터로도 처리할 수 있게 수정함.

+++ src/myproject/BpmSynchorWindow.java
@@ -264,10 +264,10 @@ public class BpmSynchorWindow {
 			public void stateChanged(ChangeEvent e) {
 				System.out.println("spinner Changed Handler.."+ spnrBpm.getValue() );
 				//				spnrBpm.getNumber();
-				waveSynchPane.setBpm( Integer.parseInt(""+spnrBpm.getValue()) );
+				waveSynchPane.setBpm( Float.parseFloat(""+spnrBpm.getValue()) );
 			}
 		});
-		spnrBpm.setModel(new SpinnerNumberModel(80, 20, 280, 1));
+		spnrBpm.setModel(new SpinnerNumberModel(80.0f, 20.0f, 280.0f, 1.0f));
 		
 		JLabel lblCategory = new JLabel("Category:");
